[BPF] disable conntrack bypass, exclude link-local
Given the fact that since kernel 5.9+ when FIB is enabled (always except
if ipip is used) we always bypass iptables for forwarded traffic, we no
longer have the confusing half-open conntrack entries because the first
packet would sometimes go through host net stack and then the rest would
bypass it. And since most deployments do not use kernels older than 5.9
it is not a common requirement to bypass conntrack. Therefore the
default setting it not to do that. However, it is still possible to turn
it on via bpfHostConntrackBypass. Some environments like MKE require
this to be turned off to function correctly.

Also link-local addresses destinations should not be excluded.

fixes https://github.com/projectcalico/calico/issues/9157