[BPF] Use bpf_redirect_peer when CALI_CT_ESTABLISHED_BYPASS
When a packet arrives on a host interface from the network and conntract
says that the status is CALI_CT_ESTABLISHED_BYPASS, that means, the
packet should be forwarded to another interface (workload) and the BPF
program on that interface would just let it in, it is faster to skip
that BPF program and redirect is straight to the ingress side of the
workload device (veth). bpf_redirect_peer() does it in the same context
in which it processes the ingress on the host interface.

This should cover almost all in-cluster traffic when CTLB is enabled
since there is no NAT happening (done in connect) and thus we can
preapprove traffic on the host device.

Note that this works for IPIP - we do not have FIB and forwarding
enabled for IPIP, but when a packet arrives to the host via IPIP, it is
fast forwarded to the destination workload as well.