Bug: csidriver crash loop / pod terminating forever (#9279)
Pods which use nodeagent volumes will stay in Terminating phase forever, if their node reboots while they are in the Terminating phase. The chances of this happening increase when a termination grace period is set on a pod whose container only responds to SIGKILL.

Post-boot, the container is not restored and we don't see its nodeagent files appear in the host FS, since it is meant to be terminating. Kube will then invoke our CSI-driver's TearDown routine. The TearDown routine fails because we cannot find the nodeagent files which we assume are mounted (even though we only want to delete them anyway...).

This commit logs "NotExists" FS errors and continues, rather than logging and exiting, to ensure we don't block the TearDown process.

Also adds UTs to the `pod2daemon` pkg and fixes a Makefile bug which made test failures exit with RC 0.