Enhance conntrack map flexibility with CPU-based scaling
Previously, BPFMapSizeConntrack was set to a fixed value, which lacked
flexibility across different spec of machines. Now, with the introduction
of BPFMapSizeConntrackWithCPUFactor, the conntrack map size can be adjusted
based on the number of CPU cores. This allows for larger map sizes on
high-spec machines and smaller map sizes on lower-spec machines, optimizing
resource usage accordingly.

BTW, recently, we added several high-spec machines to the data center, which
led to the conntrack map size being filled up on these machines. It could
have been avoided with BPFMapSizeConntrackWithCPUFactor IHMO.

Signed-off-by: Mingzhe Yang <mingzhe.yang@ly.com>
Signed-off-by: Lance Yang <ioworker0@gmail.com>