Fix List() calls that didn't fill in revision (#9599)
* Fix List() calls that didn't fill in revision

The watchercache requires the KVList to contain the revision
for correctness.  Without it, it can miss updates from the server.

Fix the resources that were missing the revision and panic if the
revision is missing to surface the problem in future.

* Fix missing rev if list turns into a get.

Replace Gets() with lists that filter on name.

* Allow for bad revision on empty list.

Looks like k8s might return that in some cases (customK8sResourceClient
checks for it).

* Add special-case list-to-get translation for CNI plugin

CNI plugin relies on its List() being turned into a Get()
for a single Pod.  It doesn't have RBAC permissions for list
and it only uses List() because one pod can result in >1 WEP.

* Add low-level tests for List calls.