[BPF] conntrack more resistant to spurious RSTs
When we see an RST on a connection, we mark the connection for cleanup
from the conntrack table. That happens immediatelly when the connection
becomes idle and the TCPResetSeen timeout expires (by default 40s). Not
a huge problem for busy connections or connections with keepalive.
Nevertheless, an attacker may taint connections and once the connections
are removed from conntrack, the endpoints need to get into
retransmittions and timeouts to close their ends.

This change removes the taint if the connection sees traffic after the
RST. First it extends the timeout to 2m and if there is still traffic,
it removes the taint completely.

This mechanism is not bullet-proof since and idle connection might still
be GCed, but it reduces the possibility of connection breakage when
there is a temporary hickup in the network. Also allows to reduce the
RSTSeen timeout.