[BPF] BPFMapSizeConntrackScaling=DoubleIfFull
When the conntrack table is running out of space, it doubles its size
automatically. It is hard to pick up a good default so we start with
something "good enough" and if it is not, then we double it and perhaps
double again if need be. The memory is not preallocated and we
periodically clean ut the table so if the usage goes down, memory should
also go down.

This mode is the new default, but it can be turned off.