[BPF] Add support for tracking conntrack metrics in Userspace mode
When BPFConntrackCleanupMode is set to BPFConntrackModeUserspace, or in
cases where the kernel does not support BTF (unfortunately, this is the
case with some of our clusters), we are missing metrics related to the
conntrack map. These metrics are obviously very useful, so let's also track
these metrics under the BPFConntrackModeUserspace setting.

Suggested-by: Tomas Hruby <tomas@tigera.io>
Signed-off-by: Mingzhe Yang <mingzhe.yang@ly.com>
Signed-off-by: Lance Yang <ioworker0@gmail.com>