Fix race when setting wireguard NAPI setting. (#9797) (#9801)
Fix a race when setting up the wireguard device.  The problem was that
we were re-loading the link from the kernel but the NAPI threading
conditional checked a cached `attrs` variable instead of checking
the attrs of the reloaded link.

As belf-and-braces, return an error if NAPI cannot be configured yet so
that the int dataplane's main loop will eventually retry.  Otherwise,
we rely on the kick from the interface monitor, and it's not 100% clear
to me that we can't miss a kick due to reconfiguring the device.