[BPF] create entry in counters map when attaching to device/hook
This makes it possible to expand the number of counters beyond what the
ebpf stack can handle. Previously, an empty/zeroed entry was created when
the ebpf program failed to get the counters and we needed to allocate
enough space on stack for the empty value. This hit the wall of the
stack size.

It is easier to create values across all the CPUs at the same time when
the programs are being attached to the hooks. In addition, the all
counters are always valid and at least zero for each device.