[BPF] treat conns with both FINs as finished in flowlogs
Flowlogs for nf/iptables consider anything in a state that has seen both
FINs as expired [1], bpf should do the same. We are not removing the
conntrack, thus we are not going to prevent retransmits in case some
packets are lost. However, from the flowlogs point of view, this
connection is done, both ends decided to finish and it is just a matter
of time that both ends will come to a mutual agreement. Therefore we can
report it is expired.

For DSR, it is a bit more tricky, but we know that at least one end is
done. We are not going to see the other FIN, so it does not matter if we
decide that it is expired immediately or wait.

[1] https://github.com/projectcalico/calico/blob/931244d78cfbb2c6384054a6b2ce06357815581c/felix/collector/iptables.go#L187-L190