Add GetAuth method to docker config
The GetAuth method can fetch authentication credentials if we have
them defined, or attempt to fetch them from the credential helper or
credential store if we don't.

If a credential helper is configured for a domain but that domain has no
entry in the auths map, we still query the credential helper anyway so
that we don't get tripped up if the auths mapping is missing entries.