OpenStack: Don't modify port dict when collecting additional information
While working on #10160 I noticed that our UT code sometimes has to reset generated elements in the mock port dicts that it uses, e.g.

    del self.osdb_ports[0]['network_qos_policy_id']

This is because our driver code gets a port dict, in a create or update hook, and then has to gather various pieces of related information, from other tables in the Neutron DB, before being ready to map that to a corresponding Calico WorkloadEndpoint; and it stores that additional information by adding keys to the supplied port dict.

It's a bit fragile to modify the input port dict like that, and it would be better to collect the additional information in a separate object, as this change now does.