[BPF] fix pkt/bytes accounting fo IPIP as L3 dev
Pretty much we need to account for the bytes on ingress from any tunnel
device. It does not matter if it preserves the marks like vxlan or ipip
or not like wg. Simply account those bytes and packets, because it is
ingress into the system and we haven't seen this as unpacked yet for
sure.

refs https://github.com/projectcalico/calico/pull/10223