[CORE-11251] Openshift: move CR order to after operator has applied CRDs
Move custom resource yamls from 01-cr-*.yaml to 03-cr-*.yaml, so that
they are applied after the operator has been created (02-*.yaml), so
that the operator is able to create the CRDs needed for the CRs.

Tested on openshift 4.16 on AWS, and HCP cluster (HCP has some docs
changes pending to correctly apply all yamls and wait for CRDs to
be created).