adjust flows expiration in collector
We need to give it more time than in iptables, better, we want to tie it
to the ScanPeriod and bpf conntrack timers. We should mostly see the
connections to expire way before they are removed from conntrack. But
just in case we should also clean up if the dataplane for any reason
stopped seding update without us seeing it as expired.