Felix: split dataplane apply time stat (#10288)
Now that we have reduced the frequency of resyncs of various kinds
the dataplane apply time stat can get heavily skewed by the start-of-day
resync.  If the cluster is very quiet, this can make it look like felix
is overloaded, instead of asleep!

Split the start-of-dat resync into its own stat.  This is useful anyway
and it avoid polluting the main stat.