Intern label KVs as unique.Handle[string]  (#10272)
* Add uniquelabels package.

Provides uniquelabels.Map, which represents resource KV
labels using unique.Handle for key and value. Implements JSON
serialisation methods to allow it to be marshalled/unmarshalled
in the same format as a normal map.  When unmarshalled, the
keys and values get deduped, saving storage space.

* Intern label KVs as unique.Handle[string]

We expect label keys and values to be reused across many endpoints
so this saves considerable amounts of storage.  Previously, every JSON
map that we decoded would have its own copies of the key and value
strings.

Furthermore, the maps themselves get smaller because each map key/value
only has to store an 8-byte handle instead of a 16-byte string. This
effect is magnified because maps often have many emptry KV slots, which
now waste only 8 bytes each instead of 16.

Of course, the interning map itself is not free, but it seems to do well
against the benchmarks in the calc package.