[BPF] Support for IPv4 fragmentation
Incoming IP fragments are stored in an LRU hash map. They can arrive out
of order. After each fragment, we check whether we have all fragments.
If any fragment is missing, we drop the skb as we cannot let it through.
Once we have all fragments, we use the current skb to assemble the whole
packet, we parse it again and we let it process by the rest of the
programs like if the packet arrived as a single chunk.

We need to defragment incoming packets because we would not be able to
pass then through policies that match on more than IP. Also we would not
be able to match them to connections in conntrack. In fact, the payload
of the fragments would be wrongly treated as L4 headers and
misinterpreted.

After a packet is reassebled, fragments are deleted. If for any reason
we never see all fragments, LRU will kick out the stored fragments
eventually.

There are some limitations:

* packet cannot have more than 10 fragments - 10 is arbitrary number
  greater than a reasonable number of fragments in modern networks (2)
  plus we fragment the packet internally into 1500 chunks in case the
  fragments were bigger than this - unlikely, but not impossible.
  However, there is no limit on fragmentation in any RFC except the
  smallest MTu of 576 bytes.
* we can store up to 10k fragments - 10k is again arbitrary. If there is
  a higher fragmentation rate than this, eBPF dataplane is probably not
  the right choice as performance would suffer and it is likely better
  to let generic Linux handle such cases.
* defragmentation is meant to handle corner cases and is not meant to be
  performant.