[BPF] track fragments from host
We need to assemble the fragments towards the host either to deal with
reordering - the first fragment with L4 headers may not arrive first -
or with NATing as it is easy to NAT a whole packet, but difficult to NAT
the first and then only do partial nating without being able to find the
CT entry due to missing L4 headers.

We assume that the host does not reorder packets and therefore we police
the first fragment, record that it is allowed and let the subsequent
fragments through. The last fragment remove the record, however, in case
of any failure and missing fragments, LRU will eventually clean it up.