feat: Add IPAM block cleanup for removed nodes
This adds a new CleanupBlocksForRemovedNodes method to the IPAM interface
and implements a new CLI command 'calicoctl ipam clean orphaned-blocks' to
help administrators clean up IPAM blocks for nodes that no longer exist.

The implementation includes:
- New interface method in libcalico-go/lib/ipam/interface.go
- Implementation in libcalico-go/lib/ipam/cleanup.go
- CLI command in calicoctl/calicoctl/commands/ipam/clean.go
- Test support in kube-controllers/pkg/controllers/node/fake_ipam_cleaner.go

This resolves the issue of slow IPAM block release when nodes are removed
from the cluster, providing a built-in alternative to external cleanup scripts.