Add support for peakrate to bandwidth QoS controls
Add support for configuring peakrate (and optionally minburst) for bandwidth QoS controls.

The following annotations are used to configure it for ingress and egress:
- "qos.projectcalico.org/ingressPeakrate"
- "qos.projectcalico.org/egressPeakrate"
- "qos.projectcalico.org/ingressMinburst"
- "qos.projectcalico.org/egressMinburst"

Minburst defaults to the interface MTU and it should be configured to no less than it (validation does set it to the MTU if configured lower).

Also improve the QoS FVs to add peakrate coverage (by verifying the first interval of the iperf3 run).