bpf_fib_lookup() may return BPF_FIB_LKUP_RET_FRAG_NEEDED
Even when mtu is OK,w we may still get BPF_FIB_LKUP_RET_FRAG_NEEDED even
when we do not ask for mtu check. The irony is that if we asked for mtu
check, we would not get BPF_FIB_LKUP_RET_FRAG_NEEDED and all would be
good.