fix regression for multihomed nodes
When there is asymetric routing on a host with multiple NICs, we have
packets returning from a different device then outbound packets are sent
to. We should not use the ingress device indef for egress forwarding.

This requires loose rpf check. We were previously switching remembered
ifindex whenever we see packets arriving from a different NIC. This
change will only remember the index if the returning packets satisfy
strict RPF - that is when the return path reflects routing.