[CORE-11447] Add support for peakrate to bandwidth QoS controls (#10458)
* [CORE-11447] Add support for peakrate to bandwidth QoS controls

Add support for configuring peakrate (and optionally minburst) for bandwidth QoS controls.

The following annotations are used to configure it for ingress and egress:
- "qos.projectcalico.org/ingressPeakrate"
- "qos.projectcalico.org/egressPeakrate"
- "qos.projectcalico.org/ingressMinburst"
- "qos.projectcalico.org/egressMinburst"

Minburst defaults to the interface MTU and it should be configured to no less than it (validation does set it to the MTU if configured lower).

Also improve the QoS FVs to add peakrate coverage (by verifying the first interval of the iperf3 run).

* Increase maximum allowed burst size to the maximum value supported by
the underlying netlink library, (max.MaxUint32 bytes, so 8 times that in
bits). Keep the default burst size at 4Gi bits (512 Mi bytes).

* make parseIperfJsonOutput more robust and add logging

* update tests

* - Address review comments
- Add comments clarifying examples and units used
- Remove unrelated Makefile line change
- Other improvements:
  - Cap to math.MaxUint32 when casting to uint32
  - Retry iperf client when json report is empty
  - Add validation for when peakrate is smaller than bandwidth

* improve comments + make generate

* use Eventually() to make packet rate FV tests more robust

* get rid of eventually for the qos FVs, use the existing retry function

* make peakrate limits higher than bandwidth limits

* update comment

* add comment

* Update libcalico-go/lib/apis/v3/workloadendpoint.go

fix typo in comment

Co-authored-by: Nell Jerram <nell@tigera.io>

* don't loop through qdiscs when a Tbf is found

* make generate

* Update QoSControls field comments

---------

Co-authored-by: Nell Jerram <nell@tigera.io>