[BPF] scale trampoline stride in policy programs
When jit_harden is enabled, some long jump may get over the allowed
distance of +/-15bit when rewritten by the kernel. When this happens, we
may retry and recompile the policy programs with the trampolines that
make the jumps shorter. If that still fails we may retry with even more
dense trampolines.

We remember the newly found stride for the given node so that we do not
need to keep figuring out for other policy programs. It is likely that
many policies would face the same issue.

Note that the hardedning already increases the amount of executed
instructions and adding some extra trampolines have probably much
smaller effect than in a non-hardened program.