Enable pre-merge packaging via a manual promotion
Sometimes we want to test a change in our Calico/OpenStack ST (`calico-test`) before merging it.
But the Calico/OpenStack ST is driven by packaged code, so this means we need to package the
pre-merge code.  The changes in this PR enable that, as follows:

1. Push the relevant code to a branch of github.com/projectcalico/calico.  It must be that repo, not
a fork, because the process needs secrets that we don't share with forks.

2. Create a PR from that branch.

3. Create a PPA at https://launchpad.net/~project-calico named `pr-<N>`, where `<N>` is the number
of the PR.  (If that PPA doesn't already exist.)

4. Click the "Publish openstack packages" promotion.

That should result in packages with the pre-merge code being uploaded to the `pr-<N>` PPA.  When
those have been built and published, you can push a branch to github.com/tigera/calico-test with,
for example:

    modified   build_and_run_fv.sh
    @@ -23,7 +23,7 @@ CURRENT_BRANCH=`git rev-parse --abbrev-ref HEAD`
     : ${BUILD_NUMBER:=0}
     : ${WORKSPACE:="$(pwd)"}
     : ${PKG_SRC_STEM:='master'}
    -: ${INSTALL_SOURCE:=""}
    +: ${INSTALL_SOURCE:="ppa:project-calico/pr-10610"}
     # PKG_SRC_STEM is only used if INSTALL_SOURCE is not set.
     : ${TEST_BRANCH:=${CURRENT_BRANCH}}
     : ${TEST_REQ:='smoke'}

and that will kick off an ST run using that PPA.