Auto resize the conntrack map from the userspace cleaner. (#10617)
* Auto resize the conntrack map from the userspace cleaner.
Userspace cleaner scans the conntrack map and takes account of
the current live entries and delete the stale entries. If it sees
that the table is 85% or more full for 3 continuous iterations/if it
sees the table is 90% full, the conntrack map size will be doubled.
This logic is already in BPF cleaner, which we are not going to use
due to LRU.

* Remove debug prints

* Auto resize ctMap only if auto-scaling is enabled.