[BPF] Set imissing tunnel key on vxlan if BYPASS is set
Workload endpoints set the vxlan tunnel key just before they redirect
the packet to the vxlan device. If for any reason redirection is
disabled, but the packet is routed via vxlan, it may be that BYPASS is
set on an established connection. The vxlan device must not
short-cirquit everything, it still needs to fill in the tunnel key as it
does for host traffic.