BPF: Fix BW QoS controls (#10694)
In order for BW QoS controls to work properly on BPF mode, there should
not be any redirection that bypasses the caliXXXX intf that has the tc
qdisc attached.

Set a global flag to skip redirection if egress BW QoS is configured
(CALI_GLOBALS_SKIP_EGRESS_REDIRECT), and set route and state
level flags (CALI_RT_SKIP_INGRESS_REDIRECT and CALI_ST_SKIP_REDIR_PEER)
to skip redirection when ingress BW QoS is configured (as the traffic
originates in other BPF programs in the ingress case). Ingress
redirection is also skipped in the case of VM workloads (e.g. kubevirt).

As such, a bit of the code that was VM workload-specific was generalized
in order to cover both VM workloads and BW QoS (and possibly other use
cases if there are any in the future). Notably, the WorkloadType field was
removed from the WorkloadEndpoint protobuf, in favor of a WorkloadBpfSkipRedir
(even though, currently only the ingress value is used).

Rename and extend the VM workload UT to cover skip ingress redirection,
both from HEPs and from WEPs.

Extend the QoS FVs to work in BPF mode, and to cover different
encapsulations (VXLAN, IPIP, no-encap). Add some skips to trim down the
number of test cases (e.g. encapsulations on iptables/nftables mode,
etcd datastore on BPF mode).