This implements a bpf conntrack cleaner. The userspace scanners (liveness and stale NAT) decide if the conntrack entry has expired. If it has expired, it returns ScanVerdictDelete along with the timestamp of the entry. 1. For Normal and Reverse entry, ScanVerdictDelete and the timestamp is returned. 2. For NATForward entry, ScanVerdictDelete and timestamp of the reverse entry is returned, if    the reverse entry is present. If reverse entry is absent, the timestamp of the forward entry    is returned.
Upon receiving ScanVerdictDelete, the conntrack key is added to the conntrack cleanup map. The
key to the map is the conntrack key and the value has the reverse key, timestamp and reverse entry's
timestamp. The reverse key is present only for NATForward entry.

	<conntrack_key> ==> <reverse_conntrack_key>:<time_last_seen>:<time_last_seen_reverse>

This is added to the conntrack cleanup map.

After all the entries are scanned, the userspace scanner calls the bpf cleaner.

BPF Cleaner
===========
The BPF cleaner, walks the conntrack cleanup map entry by entry and does the following
1. If the value, has a valid reverse key, lookup the conntrack map with the reverse key. If
   the last_seen of the conntrack value matches, the <time_last_seen_reverse>, delete
   both the reverse and forward entry from the conntrack map.

2. If the value doesn't have a valid reverse key, lookup the conntrack map with key, get the value
   and if the last_seen of the value matches the <time_last_seen> recorded in the cleanup map, delete the
   entry.