Replace IPSetMember struct with an interface (#10747)
The `IPSetMember` struct:

- Was very large, with high occupancy (we store hundreds of thousands
  of them in larger clusters).
- Contained lots of fields that were only used in a minority of cases.
- Was error prone: which fields were allowed to be "zero" and in which
  combinations?

This commit replaces the struct with an interface along with several
small implementation structs covering each use case.  This reduces
memory usage a lot and it makes it explicit which combinations of
fields are required.

The new interface moves to its own package to reduce import cycle fun.

Move the protobuf encoding to the implementation structs; this is
cleaner than a type switch in the event sequencer.

For the very common cases (simple IP-only or CIDR-only members), go
the extra mile and make variants of the structs that store:

- A single IPv4 CIDR
- A single IPv6 CIDR
- A single IPv4 addr
- A single IPv6 addr

This means that, instead of a large struct embedding an `ip.CIDR`
interface, which in turn points at a CIDR struct, we just have an
interface that points at a CIDR (or an even smaller IP) struct.

Also: remove the expectation that IP set members should be lower
case in the FV tests. This was interfering with various IP set
types in enterprise.