eBPF: Hybrid conntrack cleaner (#10736)
* This implements a bpf conntrack cleaner. The userspace scanners (liveness and stale NAT)
decide if the conntrack entry has expired. If it has expired, it returns ScanVerdictDelete
along with the timestamp of the entry.
1. For Normal and Reverse entry, ScanVerdictDelete and the timestamp is returned.
2. For NATForward entry, ScanVerdictDelete and timestamp of the reverse entry is returned, if
   the reverse entry is present. If reverse entry is absent, the timestamp of the forward entry
   is returned.

Upon receiving ScanVerdictDelete, the conntrack key is added to the conntrack cleanup map. The
key to the map is the conntrack key and the value has the reverse key, timestamp and reverse entry's
timestamp. The reverse key is present only for NATForward entry.

	<conntrack_key> ==> <reverse_conntrack_key>:<time_last_seen>:<time_last_seen_reverse>

This is added to the conntrack cleanup map.

After all the entries are scanned, the userspace scanner calls the bpf cleaner.

BPF Cleaner
===========
The BPF cleaner, walks the conntrack cleanup map entry by entry and does the following
1. If the value, has a valid reverse key, lookup the conntrack map with the reverse key. If
   the last_seen of the conntrack value matches, the <time_last_seen_reverse>, delete
   both the reverse and forward entry from the conntrack map.

2. If the value doesn't have a valid reverse key, lookup the conntrack map with key, get the value
   and if the last_seen of the value matches the <time_last_seen> recorded in the cleanup map, delete the
   entry.

* Fix format

* Update cleanup stats

* When scanning entries, queue up NAT forward and reverse entries together
and make sure they are deleted at the same time.

* Check if the fwd entry still points to the same reverse NAT entry
before cleaning up. In UTs, ensure we are not leaking any keys
in the cleanup map.

* Fallback to userspace cleaner if the bpf cleaner program fails to
load. Run the bpf cleaner for every 1000 keys for a cleanup.

* Change cleanup map to caching map.

* use interface for ipversion handling

* Reset the caching map after every run of the bpf cleaner.

* Address review comments

* Run the bpf cleaner only if there are expired entries

* Revert changes to bpf map resize FV

* Address review comment

* Using memcmp for comparing ct keys