Fix networking-calico problems reported by flake8
At the start of this exercise, flake8 (run in the networking-calico directory) reports 472 moans:

    flake8 | wc -l
    472

A lot of those are line length, because flake8's default, 79, is different from black's default of
88.  So add a .flake8 with max-line-length 88.  Then:

    flake8 | wc -l
    193

There are still some "line too long" moans that we have to fix manually.  After those the count is
down to 151.