Make OpenStack resync interval configurable and disable-able
In principle the ongoing resync isn't needed at all, so let's make it possible to switch it off, and
do our DevStack (FV) testing like that, to confirm the principle.

On the other hand, if it is wanted, let's make the period configurable.  For now we continue to
default to the same as before, i.e. 60s.

Plus a small variable naming improvement.