Use different subnet CIDR to avoid clash with following Tempest test
In the Tempest test_minimum_basic_scenario test, test code was failing to SSH to the cirros VM, and
the logs revealed this:

    Sep 05 10:03:50 semaphore-vm dnsmasq[100730]: failed to create listening socket for 192.168.100.1: Address already in use
    Sep 05 10:03:50 semaphore-vm dnsmasq[100730]: FAILED to start up
    Sep 05 10:03:50 semaphore-vm calico-dhcp-agent[83112]: ERROR neutron.agent.linux.utils [-] Exit code: 2; Cmd: ['env', 'PROCESS_TAG=dnsmasq-bbb6e5e6-772d-4043-86a6-666b10a7bb08', 'dnsmasq', '--no-hosts', '--no-resolv', '--except-interface=lo', '--pid-file=/opt/stack/data/neutron/dhcp/bbb6e5e6-772d-4043-86a6-666b10a7bb08/pid', '--dhcp-hostsfile=/opt/stack/data/neutron/dhcp/bbb6e5e6-772d-4043-86a6-666b10a7bb08/host', '--addn-hosts=/opt/stack/data/neutron/dhcp/bbb6e5e6-772d-4043-86a6-666b10a7bb08/addn_hosts', '--dhcp-optsfile=/opt/stack/data/neutron/dhcp/bbb6e5e6-772d-4043-86a6-666b10a7bb08/opts', '--dhcp-leasefile=/opt/stack/data/neutron/dhcp/bbb6e5e6-772d-4043-86a6-666b10a7bb08/leases', '--dhcp-match=set:ipxe,175', '--bind-dynamic', '--interface=ns-bbb6e5e6-77', '--dhcp-range=set:subnet-0ed39d27-31f6-478b-9a0f-0793a05dc468,192.168.100.0,static,255.255.255.0,86400s', '--dhcp-option-force=option:mtu,1500', '--dhcp-lease-max=256', '--conf-file=', '--domain=openstacklocal', '--enable-ra', '--interface=tap27be1191-9a', '--bridge-interface=ns-bbb6e5e6-77,tap27be1191-9a']; Stdin: ; Stdout: ; Stderr:
    Sep 05 10:03:50 semaphore-vm calico-dhcp-agent[83112]: dnsmasq: failed to create listening socket for 192.168.100.1: Address already in use

Very likely this is because the QoS tests already used CIDR 192.168.100.0/24 - which means our DHCP
agent configured dnsmasq to listen on 192.168.100.1 - and that that wasn't properly cleaned up at
the end of the QoS tests.  The QoS tests _do_ clean up their network, and I can see Neutron server
DELETE_SUBNET handling in the logs, but I suspect our DHCP agent code does not fully handle that.
For now let's just fix the current CI for using a different CIDR in the QoS tests.