OpenStack: add Calico task ID to help reading Neutron server logs
The execution model of the Neutron server is complex.  It runs as multiple OS processes, each of
which has only one OS thread, but each process uses eventlet to run multiple green threads in
parallel.  In the near-ish future eventlet will be removed and there will be multiple OS threads
instead.  So the calls into our driver code can be from multiple OS processes, or from multiple
threads within the same OS process, or from multiple green threads within the same OS
process/thread.  And the processing for such calls can be interleaved - e.g. one thread or green
thread yields for a while and allows others to run.  That makes it difficult to follow through logs
- for example, to tell if a given log line is part of an UPDATE_PORT_POSTCOMMIT for port A or an
earlier UPDATE_PORT_POSTCOMMIT for port B, or some other trigger into our code.

Happily OpenStack upstream has kinda solved this problem with oslo_log and oslo_context.  We can
create an instance of oslo_context.context.RequestContext (or subclass) at the start of each of our
entry points, and arrange for that to return whatever context key/value pairs we want to appear in
each log.

Unfortunately we can't add a Calico-specific key, because then all the non-Calico Neutron logs (but
within the same task) spew a traceback.  So instead we add Calico context info into a key that
Neutron is already using.  And we also want to use a key that is already included in the default
logging formats (`logging_context_format_string` and `logging_default_format_string`).  Putting all
that together, it seems our best option is the piggy-back on the `request_id` key.

Also, when trying to follow Neutron server logs, I've found it's made harder by superfluous "Entity
rule" lines, which only contain information that is logged again a few lines later.  For example:

    2025-09-11 09:37:56.230 55140 DEBUG networking_calico.plugins.ml2.drivers.calico.policy [req-2523ac64-aae3-4ed3-ae3b-dec3a1f25555 req-96b587e8-1a34-42ea-8a9c-e02de3f071fe dea4263e8ce24fcdbc8cc2599abaf1cb 8b0c1e4b15884769b3ec1dd91e65f9e7 - - default default] => Entity rule {'selector': 'has(sg.projectcalico.org/openstack-08b351c1-0a4e-4596-ab71-7bafefe606ae)'} _neutron_rule_to_etcd_rule /usr/lib/python3/dist-packages/networking_calico/plugins/ml2/drivers/calico/policy.py:202
    2025-09-11 09:37:56.231 55140 DEBUG networking_calico.plugins.ml2.drivers.calico.policy [req-2523ac64-aae3-4ed3-ae3b-dec3a1f25555 req-96b587e8-1a34-42ea-8a9c-e02de3f071fe dea4263e8ce24fcdbc8cc2599abaf1cb 8b0c1e4b15884769b3ec1dd91e65f9e7 - - default default] => ingress Calico rule {'action': 'Allow', 'ipVersion': 6, 'source': {'selector': 'has(sg.projectcalico.org/openstack-08b351c1-0a4e-4596-ab71-7bafefe606ae)'}} _neutron_rule_to_etcd_rule /usr/lib/python3/dist-packages/networking_calico/plugins/ml2/drivers/calico/policy.py:217

So let's remove those.